% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amap_audit.R
\name{amap_audit_enable}
\alias{amap_audit_enable}
\title{Enable audit logging for amapGeocode}
\usage{
amap_audit_enable(path)
}
\arguments{
\item{path}{Required.
Directory where `amap_audit.jsonl` will be written.}
}
\description{
Enable audit logging for amapGeocode
}
