% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amap_http.R
\name{amap_sign}
\alias{amap_sign}
\title{Generate Amap signature}
\usage{
amap_sign(params, secret, path)
}
\arguments{
\item{params}{Required.
Named list of request parameters to sign.}

\item{secret}{Required.
Secret key used for request signing.}

\item{path}{Required.
Request path portion of the API URL.}
}
\description{
Generate Amap signature
}
