% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapGeocode-package.R
\docType{package}
\name{amapGeocode-package}
\alias{amapGeocode}
\alias{amapGeocode-package}
\title{amapGeocode: An Interface to the 'AutoNavi Maps' API Geocoding Services}
\description{
Getting and parsing data of location geocode/reverse-geocode and administrative regions from 'AutoNavi Maps'\url{https://lbs.amap.com/api/webservice/summary} API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://womeimingzi11.github.io/amapGeocode}
  \item \url{https://github.com/womeimingzi11/amapGeocode}
  \item Report bugs at \url{https://github.com/womeimingzi11/amapGeocode/issues}
}

}
\author{
\strong{Maintainer}: Han Chen \email{chenhan28@gmail.com}

Other contributors:
\itemize{
  \item Wanyanhan Jiang \email{243016042@qq.com} [contributor]
}

}
\keyword{internal}
