% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoord.R
\name{extractConvertCood}
\alias{extractConvertCood}
\title{Extract converted coordinate points from convertCoord request}
\usage{
extractConvertCood(res)
}
\arguments{
\item{res}{Required.\cr
Response from convertCoord.}
}
\value{
Returns a tibble which extracts converted coordinate points from request of convertCoord. See \url{https://lbs.amap.com/api/webservice/guide/api/convert} for more information.
}
\description{
Extract converted coordinate points from convertCoord request
}
