% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_admin.R
\name{mod_admin_server}
\alias{mod_admin_server}
\title{Admin Region Module Server}
\usage{
mod_admin_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Reactive expression containing the admin data
}
\description{
Admin Region Module Server
}
