% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_convert_server}
\alias{mod_convert_server}
\title{Coordinate Conversion Module Server}
\usage{
mod_convert_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Reactive expression containing the results
}
\description{
Coordinate Conversion Module Server
}
