% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geocode.R
\name{mod_geocode_server}
\alias{mod_geocode_server}
\title{Geocoding Module Server}
\usage{
mod_geocode_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Reactive expression containing the geocoded data
}
\description{
Geocoding Module Server
}
