% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_server}
\alias{mod_map_server}
\title{Interactive Map Module Server}
\usage{
mod_map_server(id, data_reactive)
}
\arguments{
\item{id}{Module ID}

\item{data_reactive}{Reactive expression returning a data frame with 'lng' and 'lat' columns.}
}
\description{
Interactive Map Module Server
}
