% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_regeo.R
\name{mod_regeo_server}
\alias{mod_regeo_server}
\title{Reverse Geocoding Module Server}
\usage{
mod_regeo_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Reactive expression containing the results
}
\description{
Reverse Geocoding Module Server
}
