% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num_coord_to_str_loc}
\alias{num_coord_to_str_loc}
\title{Take longitude and latitude from location string out.}
\usage{
num_coord_to_str_loc(lng, lat)
}
\arguments{
\item{lng}{Required. \cr
Longitude in decimal}

\item{lat}{Required. \cr
Latitude in decimal}
}
\value{
Comma binded coordinate string
}
\description{
Take longitude and latitude from location string out.
}
