% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_loc_to_num_coord}
\alias{str_loc_to_num_coord}
\title{Take longitude and latitude from location string out.}
\usage{
str_loc_to_num_coord(str_location)
}
\arguments{
\item{str_location}{Required. \cr
Location string from response}
}
\value{
vector contains Longitude and Latitude in numeric
}
\description{
Take longitude and latitude from location string out.
}
