% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amap_http.R
\name{with_amap_signature}
\alias{with_amap_signature}
\title{Execute code with temporary signature settings}
\usage{
with_amap_signature(secret, expr, key = NULL, enabled = TRUE)
}
\description{
Execute code with temporary signature settings
}
