% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoord.R
\name{convertCoord}
\alias{convertCoord}
\title{Convert coordinate from different coordinate systems to AutoNavi system}
\usage{
convertCoord(
  locations,
  key = NULL,
  coordsys = NULL,
  sig = NULL,
  output = NULL,
  to_table = TRUE
)
}
\arguments{
\item{locations}{Required. \cr
String coordinate point from other coordinate system}

\item{key}{Optional.\cr
Amap Key. \cr
Applied from AutoNavi Map API official website\url{https://lbs.amap.com/dev/}}

\item{coordsys}{Optional.\cr
Coordinate System. \cr
Support: `gps`,`mapbar`,`baidu` and `autonavi`-not convert}

\item{sig}{Optional.\cr
Digital Signature.\cr
How to use this argument? Please check here{https://lbs.amap.com/faq/account/key/72}}

\item{output}{Optional.\cr
 Output Data Structure. \cr
Support JSON and XML. The default value is JSON.}

\item{to_table}{Optional.\cr
Transform response content to tibble.\cr#'}
}
\value{
Returns a JSON, XML or Tibble of results containing detailed geocode information. See \url{https://lbs.amap.com/api/webservice/guide/api/convert} for more information.
}
\description{
Convert coordinate from different coordinate systems to AutoNavi system
}
