% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmin.R
\name{extractAdmin}
\alias{extractAdmin}
\title{Get Subordinate Administrative Region from getAdmin request
Now, it only support extract the first layer of subordinate administrative region information.}
\usage{
extractAdmin(res)
}
\arguments{
\item{res}{Response from getAdmin.}
}
\value{
Returns a tibble which extracts detailed subordinate administrative region information from results of getCoord. See \url{https://lbs.amap.com/api/webservice/guide/api/district} for more information.
}
\description{
Get Subordinate Administrative Region from getAdmin request
Now, it only support extract the first layer of subordinate administrative region information.
}
\examples{
\dontrun{
library(dplyr)
library(amapGeocode)

# Set the amap_key which is applied from 'AutoNavi' Map Services for amapGeocde globally.
options(amap_key = 'REPLACE THIS BY YOUR KEY')

#Get subordinate administrative regions as a XML
getAdmin('Sichuan Province', output = 'XML') \%>\%
   # extract subordinate administrative regions as a tibble
   extractAdmin()
}

}
\seealso{
\code{\link{getAdmin}}
}
