% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoord.R
\name{extractCoord}
\alias{extractCoord}
\title{Extract geocoding}
\usage{
extractCoord(res)
}
\arguments{
\item{res.}{Required.\cr
Response from getCoord.}
}
\value{
Returns a tibble which extracts detailed geocode information from results of getCoord. See \url{https://lbs.amap.com/api/webservice/guide/api/georegeo} for more information.
}
\description{
Extract geocoding result from Response of getCoord. For now, only single place response is supported.
}
