% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocal.R
\name{extractLocation}
\alias{extractLocation}
\title{Extract location from coordinate request}
\usage{
extractLocation(res)
}
\arguments{
\item{res}{Required.\cr
Response from getLocation.}
}
\value{
Returns a tibble which extracts detailed location information from results of getLocation. See \url{https://lbs.amap.com/api/webservice/guide/api/georegeo} for more information.
}
\description{
Extract location result from Response of getCoord. For now, only single place response is supported.
}
