% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{extractLocation}
\alias{extractLocation}
\title{Extract location from coordinate request}
\usage{
extractLocation(res)
}
\arguments{
\item{res}{Required.\cr
Response from getLocation.}
}
\value{
Returns a tibble which extracts detailed location information from results of getLocation. See \url{https://lbs.amap.com/api/webservice/guide/api/georegeo} for more information.
}
\description{
Extract location from coordinate request
}
\examples{
\dontrun{
library(dplyr)
library(amapGeocode)

# Set the amap_key which is applied from 'AutoNavi' Map Services for amapGeocde globally.
options(amap_key = 'REPLACE THIS BY YOUR KEY')
# Get reverse-geocode as a XML
getLocation(104.043284, 30.666864, output = 'XML') \%>\%
   # extract reverse-geocode regions as a tibble
   extractLocation()
}

}
\seealso{
\code{\link{getLocation}}
}
